requirements-setup.txt
# Dependencies for setup_repo.py with secret management
PyGithub>=1.59.0
PyYAML>=6.0
requests>=2.31.0
PyNaCl>=1.5.0
